requirements.txt
pydantic>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
requests>=2.31.0
prefect>=2.14.0
prefect-sqlalchemy>=0.4.0 